+++ lib/templates/versioning/downloadableTemplateVersion.ts
@@ -15,7 +15,6 @@ export type DownloadableTemplateVersion = {
   label: DownloadableTemplateVersionLabel;
   formConfig?: FormProperties;
 };
-
 export type DownloadableTemplateVersionsInput = {
   currentDraftVersion?: { id: string; versionNumber: number } | null;
   currentPublishedVersion?: { id: string; versionNumber: number } | null;
@@ -23,50 +22,3 @@ export type DownloadableTemplateVersionsInput = {
   currentPublishedVersionId?: string | null;
   versions?: Array<{ id: string; versionNumber: number }> | null;
 };
-
-export function formatDownloadableTemplateVersions(
-  template: DownloadableTemplateVersionsInput
-): DownloadableTemplateVersion[] {
-  const downloadableVersions: DownloadableTemplateVersion[] = [];
-
-  if (template.currentDraftVersion) {
-    downloadableVersions.push({
-      id: template.currentDraftVersion.id,
-      versionNumber: template.currentDraftVersion.versionNumber,
-      label: DOWNLOADABLE_TEMPLATE_VERSION_LABEL.currentDraft,
-    });
-  }
-
-  if (template.currentPublishedVersion) {
-    downloadableVersions.push({
-      id: template.currentPublishedVersion.id,
-      versionNumber: template.currentPublishedVersion.versionNumber,
-      label: DOWNLOADABLE_TEMPLATE_VERSION_LABEL.currentPublished,
-    });
-  }
-
-  const skippedVersionIds = new Set(
-    [template.currentDraftVersionId, template.currentPublishedVersionId].filter(Boolean)
-  );
-
-  (template.versions || []).forEach((version) => {
-    if (skippedVersionIds.has(version.id)) {
-      return;
-    }
-
-    downloadableVersions.push({
-      id: version.id,
-      versionNumber: version.versionNumber,
-      label: DOWNLOADABLE_TEMPLATE_VERSION_LABEL.published,
-    });
-  });
-
-  if (downloadableVersions.length === 0) {
-    downloadableVersions.push({
-      versionNumber: 1,
-      label: DOWNLOADABLE_TEMPLATE_VERSION_LABEL.currentDraft,
-    });
-  }
-
-  return downloadableVersions.sort((left, right) => right.versionNumber - left.versionNumber);
-}
